Financial-Planner
This financial-planner.jpynb (jupyter notebook) file includes the personal finance planner application, and financial health analysis comprising part 1. Part 2 of the exercise focuses on the Retirement planning section over a 30 year period as well as Monte Carlo simulations using the Alpaca API to fetch historical data and a Retirement Analysis synopsis. It also includes the Optional Challenge whereby we adjust the portfolio to have a higher initial investment and rerun the retirement analysis to see what it would take to retire in `5` or `10` years instead of `30`.